1週間単位のトレーニング記録を取得する機能 フロントでのメニュー、回数、セット数等をDBにinsert完了

+++ src/main/java/com/example/springbootdemo/controllers/training/requests/CreateTrainingRequest.java
@@ -10,16 +10,16 @@ import java.time.*;
 public class CreateTrainingRequest {
 
     @NotNull
-    private Long id;
+    private int id;
 
     @NotNull
     private String trainingMenu;
 
     @NotNull
-    private Long times;
+    private int times;
 
     @NotNull
-    private Long reps;
+    private int reps;
 
     @NotNull
     private LocalDate date;
